requirements.txt
streamlit
python-dotenv
pandas
numpy
python-docx
python-pptx
PyPDF2
openpyxl
langchain
langchain-experimental
faiss-cpu
openai
anthropic
plotly
SQLAlchemy
tqdm
requests
tenacity
tiktoken
pdfplumber
pytesseract